backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Security and authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Database drivers
motor>=3.3.0  # MongoDB async driver

# File handling
aiofiles>=23.2.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
numpy>=1.24.0

# Additional utilities
python-dotenv>=1.0.0

# OCR / Vision
opencv-python>=4.6.0.66
pillow>=10.1.0
pytesseract>=0.3.10
paddleocr>=2.7.0.3
pymupdf>=1.24.0

# NLP / AI
spacy>=3.7.2

# Export
openpyxl>=3.1.2

